requirements.txt
pyquery
requests
python-frontmatter==1.0.0
pytz==2024.1
